requirements.txt
streamlit>=1.30.0
google-generativeai>=0.3.0
SpeechRecognition>=3.10.0
pydub>=0.25.1
google-cloud-texttospeech>=2.16.0
streamlit-audiorecorder>=0.1.2
